services/logging/requirements.txt
flask==3.0.3
paho-mqtt==2.1.0
docker==7.1.0
